% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deactivate.r
\name{deactivate}
\alias{deactivate}
\title{Deactivate verse}
\usage{
deactivate(...)
}
\arguments{
\item{...}{Unused arguments, for future development.}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
Used to deactivate verse. The \code{.Rprofile} file is deleted from
the top level of the project directory and the project library is removed
as a library path.
}
