% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lock.r
\name{._read_lock}
\alias{._read_lock}
\title{Read and parse a \code{verse.lock} file.}
\usage{
._read_lock(...)
}
\arguments{
\item{...}{Unused arguments, for future development.}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
Read and parse a \code{verse.lock} file.
}
