% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse_install.r
\name{._verse_install_dep}
\alias{._verse_install_dep}
\title{Install package dependencies from CRAN}
\usage{
._verse_install_dep(...)
}
\arguments{
\item{...}{Unused arguments, for future development.}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
The function identifies the dependencies of all packages within
the project library. If the project library doesn't contain a dependency
then it is installed and added to the \code{verse.lock} file.

Note that dependencies in this instance refers to R packages and not system
or other dependencies. Base packages are never installed in the project
library.
}
