% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollback_lock.r
\name{rollback_lock}
\alias{rollback_lock}
\title{Rollback a \code{verse.lock} file to a previous state.}
\usage{
rollback_lock(...)
}
\arguments{
\item{...}{Unused arguments, for future development.}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
When a change is made to the lock file the prior configuration
is stored whilst the session remains active. This allows a brief opportunity
to test code after updating or installing a package and rollback the lock
file if problems are identified.
}
\details{
The temporary storing of the prior configuration of the lock file
only persists until the session is terminated as is intended only as a short
term feature with limited usage.
}
