% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_lock.r
\name{verify_lock}
\alias{verify_lock}
\title{Verify the project library}
\usage{
verify_lock(...)
}
\arguments{
\item{...}{Unused arguments, for future development.}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
Used to verify that the contents of the project library adhere
to the conditions set out in the lock file.
}
\details{
Packages within the private library are checked for compliance with
the lock file. Missing packages are installed and packages whose
installations fail to meet the lock file specification are re-installed.

The function does NOT remove packages that are installed in the project
library and not contained within the lock file.
}
