% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse_install.r
\name{verse_install}
\alias{verse_install}
\title{Install a package from CRAN}
\usage{
verse_install(
  package,
  version = NULL,
  upgrade = TRUE,
  dependencies = TRUE,
  ...
)
}
\arguments{
\item{package}{string, the name of a package to install.}

\item{version}{string, specifying the version of the package to install.}

\item{upgrade}{logical, default \code{TRUE} upgrades dependencies of the specified
package where a version of the dependency is already installed.}

\item{dependencies}{logical, default \code{TRUE} installs dependencies of the
specified package.}

\item{...}{Arguments to be passed on to \code{\link[remotes:install_version]{remotes::install_version()}}}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
The function acts as a wrapper for \code{remotes::install_version()}.
It allows for the installation of a specific version of a package from CRAN.
The installed package is added to the \code{verse.lock} file.
}
\details{
Archived packages on CRAN may need to be built from source. As such
the development environment must contain the tools required to perform the
operation. Users can check whether their system is able to build packages
from source using the \code{has_devel()} function from the \code{devtools} package.

The lock file always records a specific version of a package, and future
restoration or rebuilding of the project ensures that those specific
versions are installed.

Following installation of a package the dependencies of the installed
package are checked to ensure their installation has taken place. NOTE that
in this case dependencies means only other R packages and other system
dependencies are not considered.
}
