% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lock.r
\name{write_lock}
\alias{write_lock}
\title{Write a \code{verse.lock} file.}
\usage{
write_lock(...)
}
\arguments{
\item{...}{Unused arguments, for future development.}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
Record the current state of the project library to the lock
file.
}
