% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{init}
\alias{init}
\title{Initialize verse}
\usage{
init(...)
}
\arguments{
\item{...}{Unused arguments, for future development.}
}
\value{
Invisible, the function is called for it's side effects.
}
\description{
Used to initialize verse in the working directory which should
be the top level directory of a project. Initializing creates a custom
\code{.Rprofile} and a private library within the project directory.
}
\details{
Initialization also activates verse. Once activated packages are
installed in the project library using \code{verse::install()}.

To help avoid unintended consequences verse cannot be initialized if
the directory already contains a \code{.Rprofile} or \code{verse.lock} file. If either
of these are present they must be manually removed by the user before
initialization.
}
